Split and align XML typings
